+++ src/main/java/duke/Duke.java
@@ -1,5 +1,7 @@
-import java.util.ArrayList;
-import java.util.List;
+package duke;
+
+import myTasks.TaskList;
+
 import java.util.Scanner;
 
 public class Duke {
