+++ src/main/java/myTasks/Task.java
@@ -1,6 +1,8 @@
+package myTasks;
+
 public class Task {
-    protected String description;
-    protected boolean isDone;
+    public String description;
+    public boolean isDone;
 
     public Task(String description) {
         this.description = description;
